Fix functional tests for appveyor
When executed against Windows, the test fixtures contain a mix of
escaped carriage returns and unescaped carriage returns, both of which
must be normalized. Further, different newlines between actual and
expected values result in different build hashes for the dist
configuration; therefore the tests are updated to account for this as
well.